feat(telex): enhance advanced Telex input method and configuration
- Updated the advanced Telex input method to support macOS UI exposure, persistence, and selection.
- Modified configuration format to include `telex_advanced` as a valid input method.
- Improved input method handling in the Funput composer and input controller.
- Added new settings synchronization and application logic for managing input methods.
- Enhanced documentation to reflect changes in input method options and behavior.
- Introduced new utility functions for managing shortcuts and excluded apps.
