+++ platforms/macos/Funput/Model/AppSettings.swift
@@ -1,23 +1,6 @@
 import Foundation
 import Observation
 
-/// An app where Funput stays out of the way: while it's the focused client, keys
-/// pass straight through (English), regardless of the VI/EN toggle. Matched by the
-/// client's bundle identifier; `name` is only for display in Settings.
-struct ExcludedApp: Codable, Identifiable, Hashable {
-    let id: String
-    let name: String
-}
-
-/// A text-expansion shortcut (gõ tắt): typing `trigger` then a word boundary injects
-/// `expansion` (`vn` → `Việt Nam`). `id` is a stable UUID — not the trigger — so a row
-/// can be edited inline while its trigger is briefly empty or duplicated.
-struct TextShortcut: Codable, Identifiable, Hashable {
-    var id = UUID()
-    var trigger: String
-    var expansion: String
-}
-
 /// User preferences for Funput, persisted in `UserDefaults`. The Settings UI and
 /// (later) the `IMKInputController` live in the same process, so they share this
 /// store directly. `@Observable` drives live SwiftUI updates.
@@ -113,7 +96,7 @@ final class AppSettings {
             Keys.showMenuBarIcon: true,
             Keys.vietnameseEnabled: true,
         ])
-        inputMethod = InputMethod(rawValue: defaults.integer(forKey: Keys.inputMethod)) ?? .telex
+        inputMethod = InputMethod.persisted(defaults.object(forKey: Keys.inputMethod))
         toneStyle = ToneStyle(rawValue: defaults.integer(forKey: Keys.toneStyle)) ?? .traditional
         vietnameseEnabled = defaults.bool(forKey: Keys.vietnameseEnabled)
         smartEnglishRestore = defaults.bool(forKey: Keys.smartEnglishRestore)
@@ -131,53 +114,6 @@ final class AppSettings {
             .flatMap { try? JSONDecoder().decode([ExcludedApp].self, from: $0) } ?? []
         shortcuts = defaults.data(forKey: Keys.shortcuts)
             .flatMap { try? JSONDecoder().decode([TextShortcut].self, from: $0) } ?? []
-    }
-
-    // MARK: - Excluded apps
-
-    /// Whether the focused client's app is on the exclusion list.
-    func isExcluded(_ bundleId: String?) -> Bool {
-        guard let bundleId, !bundleId.isEmpty else { return false }
-        return excludedApps.contains { $0.id == bundleId }
-    }
-
-    /// Add an app to the exclusion list (no-op if already present).
-    func addExcludedApp(_ app: ExcludedApp) {
-        guard !excludedApps.contains(where: { $0.id == app.id }) else { return }
-        excludedApps.append(app)
-    }
-
-    /// Remove an app from the exclusion list by bundle identifier.
-    func removeExcludedApp(_ id: String) {
-        excludedApps.removeAll { $0.id == id }
-    }
-
-    // MARK: - Shortcuts (gõ tắt)
-
-    /// Append an empty shortcut for the user to fill in inline.
-    func addShortcut() {
-        shortcuts.append(TextShortcut(trigger: "", expansion: ""))
-    }
-
-    /// Remove a shortcut by its row id.
-    func removeShortcut(_ id: UUID) {
-        shortcuts.removeAll { $0.id == id }
-    }
-
-    private enum Keys {
-        static let inputMethod = "inputMethod"
-        static let toneStyle = "toneStyle"
-        static let vietnameseEnabled = "vietnameseEnabled"
-        static let smartEnglishRestore = "smartEnglishRestore"
-        static let eagerRestore = "eagerRestore"
-        static let spellCheckEnabled = "spellCheckEnabled"
-        static let autoCapitalizeEnabled = "autoCapitalizeEnabled"
-        static let toggleShortcut = "toggleShortcut"
-        static let flipShortcut = "flipShortcut"
-        static let launchAtLogin = "launchAtLogin"
-        static let showMenuBarIcon = "showMenuBarIcon"
-        static let hasCompletedOnboarding = "hasCompletedOnboarding"
-        static let excludedApps = "excludedApps"
-        static let shortcuts = "shortcuts"
+        defaults.set(inputMethod.rawValue, forKey: Keys.inputMethod)
     }
 }
