+++ platforms/macos/Funput/Model/AppSettingsKeys.swift
@@ -1,0 +1,18 @@
+extension AppSettings {
+    enum Keys {
+        static let inputMethod = "inputMethod"
+        static let toneStyle = "toneStyle"
+        static let vietnameseEnabled = "vietnameseEnabled"
+        static let smartEnglishRestore = "smartEnglishRestore"
+        static let eagerRestore = "eagerRestore"
+        static let spellCheckEnabled = "spellCheckEnabled"
+        static let autoCapitalizeEnabled = "autoCapitalizeEnabled"
+        static let toggleShortcut = "toggleShortcut"
+        static let flipShortcut = "flipShortcut"
+        static let launchAtLogin = "launchAtLogin"
+        static let showMenuBarIcon = "showMenuBarIcon"
+        static let hasCompletedOnboarding = "hasCompletedOnboarding"
+        static let excludedApps = "excludedApps"
+        static let shortcuts = "shortcuts"
+    }
+}
